% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrapping estimation}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{An integer which is the number of bootstrap iterations and default is 10,000.}

\item{x}{A numeric vector.}

\item{fun}{A character string.}

\item{alpha}{A numeric value of the significance level for the confidence interval.}

\item{cx}{A numeric value of character expansion for text annotations on the histogram.}

\item{...}{Additional parameters to be passed to the histogram function.}
}
\value{
A list containing:
  \item{ci}{A numeric vector of length 2 indicating the lower and upper bounds of the confidence interval.}
  \item{fun}{The function used for the statistic.}
  \item{x}{The original sample.}
  \item{xstat}{A numeric vector of bootstrap statistics.}
}
\description{
Bootstrapping estimation
}
\examples{
set.seed(123)
sample_data <- rnorm(100)
results <- myboot2(x=sample_data, fun="mean")
print(results$ci)

}
