% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Histogram of Sample Mean from Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{A integer of the sample size for each iteration.}

\item{iter}{A integer of the number of iterations or samples to be taken.}

\item{lambda}{A numeric value of the lambda parameter for the Poisson distribution, default is 10.}

\item{...}{Additional graphical parameters to be passed to the hist() function.}
}
\value{
This function returns a layout of 3 plots:
1. A histogram of sample means with an overlaid normal distribution curve.
2. A barplot of relative frequencies of sampled y values.
3. A plot representing the probability function for the Poisson distribution.
No return value to the console.
}
\description{
This function samples from a Poisson distribution, computes the sample means,
and produces a histogram of these sample means. Additionally, it overlays a
normal distribution curve on the histogram based on the properties of the Poisson distribution.
}
\examples{
mycltp(n=2, iter=10000, lambda = 4)
}
