% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{Number of iteration}

\item{N}{Total number of items in the population}

\item{r}{Number of success}

\item{n}{Sample size}
}
\value{
The proportions of successes
}
\description{
Hypergeometric
}
\examples{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
