% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum Likelihood Estimation for Normal Distributions}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{A numeric vector representing the sample data.}

\item{mu}{A numeric vector of values for the mean (mu) parameter to be tested.}

\item{sig}{A numeric vector of values for the standard deviation (sigma) parameter to be tested.}

\item{...}{Additional arguments to be passed to the contour plotting function.}
}
\value{
A list containing:
  \item{x}{The original sample data vector.}
  \item{coord}{The coordinates of the maximum likelihood estimates in the grid.}
  \item{maxl}{The maximum value of the likelihood function.}
}
\description{
Maximum Likelihood Estimation for Normal Distributions
}
\examples{
mymlnorm(x=c(10, 12, 13, 15, 12, 11, 10),
         mu=seq(10, 15, length=1000),
         sig=seq(0.1, 4, length=1000),
         lwd=2, labcex=1)
}
