% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot of density}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{is the mean value.}

\item{sigma}{is the standard deviation.}

\item{a}{is the upper limit for shading and annotation.}
}
\value{
plot the graph with shaded area.
}
\description{
Plot of density
}
\examples{
myncurve(mu = 10, sigma = 4, a = 8)

}
