% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadmodel.R
\name{quad}
\alias{quad}
\title{Lab 4 Quadratic equation}
\usage{
quad(x)
}
\arguments{
\item{x}{A quantitative vector}
}
\value{
A vector of quadratic model y hat value
}
\description{
Lab 4 Quadratic equation
}
\examples{
quad(1:10)
}
